+++ src/modules/graphics/graphics.c
@@ -168,8 +168,8 @@ static void onCloseWindow(void) {
 static void onResizeWindow(int width, int height) {
   state.width = width;
   state.height = height;
-  state.defaultCanvas->width = width;
-  state.defaultCanvas->height = height;
+  lovrCanvasSetWidth(state.defaultCanvas, width);
+  lovrCanvasSetHeight(state.defaultCanvas, height);
 }
 
 static void* lovrGraphicsMapBuffer(StreamType type, uint32_t count) {
@@ -303,13 +303,13 @@ void lovrGraphicsSetCamera(Camera* camera, bool clear) {
     mat4_perspective(state.camera.projection[0], .01f, 100.f, 67.f * (float) M_PI / 180.f, (float) state.width / state.height);
     mat4_perspective(state.camera.projection[1], .01f, 100.f, 67.f * (float) M_PI / 180.f, (float) state.width / state.height);
     state.camera.canvas = state.defaultCanvas;
-    state.camera.canvas->flags.stereo = false;
+    lovrCanvasSetStereo(state.camera.canvas, false);
   } else {
     state.camera = *camera;
 
     if (!state.camera.canvas) {
       state.camera.canvas = state.defaultCanvas;
-      state.camera.canvas->flags.stereo = camera->stereo;
+    lovrCanvasSetStereo(state.camera.canvas, camera->stereo);
     }
   }
