+++ src/modules/graphics/opengl.c
@@ -65,6 +65,21 @@ struct Texture {
   uint8_t incoherent;
 };
 
+struct Canvas {
+  uint32_t framebuffer;
+  uint32_t resolveBuffer;
+  uint32_t depthBuffer;
+  uint32_t width;
+  uint32_t height;
+  CanvasFlags flags;
+  Attachment attachments[MAX_CANVAS_ATTACHMENTS];
+  Attachment depth;
+  uint32_t attachmentCount;
+  bool needsAttach;
+  bool needsResolve;
+  bool immortal;
+};
+
 typedef enum {
   BARRIER_BLOCK,
   BARRIER_UNIFORM_TEXTURE,
@@ -1765,7 +1780,8 @@ void lovrTextureSetWrap(Texture* texture, TextureWrap wrap) {
 
 // Canvas
 
-Canvas* lovrCanvasInit(Canvas* canvas, uint32_t width, uint32_t height, CanvasFlags flags) {
+Canvas* lovrCanvasCreate(uint32_t width, uint32_t height, CanvasFlags flags) {
+  Canvas* canvas = lovrAlloc(Canvas);
   if (flags.stereo && state.singlepass != MULTIVIEW) {
     width *= 2;
   }
@@ -1809,7 +1825,8 @@ Canvas* lovrCanvasInit(Canvas* canvas, uint32_t width, uint32_t height, CanvasFl
   return canvas;
 }
 
-Canvas* lovrCanvasInitFromHandle(Canvas* canvas, uint32_t width, uint32_t height, CanvasFlags flags, uint32_t framebuffer, uint32_t depthBuffer, uint32_t resolveBuffer, uint32_t attachmentCount, bool immortal) {
+Canvas* lovrCanvasCreateFromHandle(uint32_t width, uint32_t height, CanvasFlags flags, uint32_t framebuffer, uint32_t depthBuffer, uint32_t resolveBuffer, uint32_t attachmentCount, bool immortal) {
+  Canvas* canvas = lovrAlloc(Canvas);
   canvas->framebuffer = framebuffer;
   canvas->depthBuffer = depthBuffer;
   canvas->resolveBuffer = resolveBuffer;
@@ -1902,6 +1919,81 @@ TextureData* lovrCanvasNewTextureData(Canvas* canvas, uint32_t index) {
   }
 
   return textureData;
+}
+
+const Attachment* lovrCanvasGetAttachments(Canvas* canvas, uint32_t* count) {
+  if (count) *count = canvas->attachmentCount;
+  return canvas->attachments;
+}
+
+void lovrCanvasSetAttachments(Canvas* canvas, Attachment* attachments, uint32_t count) {
+  lovrAssert(count > 0, "A Canvas must have at least one attached Texture");
+  lovrAssert(count <= MAX_CANVAS_ATTACHMENTS, "Only %d textures can be attached to a Canvas, got %d\n", MAX_CANVAS_ATTACHMENTS, count);
+
+  if (!canvas->needsAttach && count == canvas->attachmentCount && !memcmp(canvas->attachments, attachments, count * sizeof(Attachment))) {
+    return;
+  }
+
+  lovrGraphicsFlushCanvas(canvas);
+
+  for (uint32_t i = 0; i < count; i++) {
+    Texture* texture = attachments[i].texture;
+    uint32_t slice = attachments[i].slice;
+    uint32_t level = attachments[i].level;
+    uint32_t width = lovrTextureGetWidth(texture, level);
+    uint32_t height = lovrTextureGetHeight(texture, level);
+    uint32_t depth = lovrTextureGetDepth(texture, level);
+    uint32_t mipmaps = lovrTextureGetMipmapCount(texture);
+    bool hasDepthBuffer = canvas->flags.depth.enabled;
+    lovrAssert(slice < depth, "Invalid attachment slice (Texture has %d, got %d)", depth, slice + 1);
+    lovrAssert(level < mipmaps, "Invalid attachment mipmap level (Texture has %d, got %d)", mipmaps, level + 1);
+    lovrAssert(!hasDepthBuffer || width == canvas->width, "Texture width of %d does not match Canvas width (%d)", width, canvas->width);
+    lovrAssert(!hasDepthBuffer || height == canvas->height, "Texture height of %d does not match Canvas height (%d)", height, canvas->height);
+#ifndef __ANDROID__ // On multiview canvases, the multisample settings can be different
+    lovrAssert(lovrTextureGetMSAA(texture) == canvas->flags.msaa, "Texture MSAA does not match Canvas MSAA");
+#endif
+    lovrRetain(texture);
+  }
+
+  for (uint32_t i = 0; i < canvas->attachmentCount; i++) {
+    lovrRelease(Texture, canvas->attachments[i].texture);
+  }
+
+  memcpy(canvas->attachments, attachments, count * sizeof(Attachment));
+  canvas->attachmentCount = count;
+  canvas->needsAttach = true;
+}
+
+bool lovrCanvasIsStereo(Canvas* canvas) {
+  return canvas->flags.stereo;
+}
+
+void lovrCanvasSetStereo(Canvas* canvas, bool stereo) {
+  canvas->flags.stereo = stereo;
+}
+
+uint32_t lovrCanvasGetWidth(Canvas* canvas) {
+  return canvas->width;
+}
+
+uint32_t lovrCanvasGetHeight(Canvas* canvas) {
+  return canvas->height;
+}
+
+void lovrCanvasSetWidth(Canvas* canvas, uint32_t width) {
+  canvas->width = width;
+}
+
+void lovrCanvasSetHeight(Canvas* canvas, uint32_t height) {
+  canvas->height = height;
+}
+
+uint32_t lovrCanvasGetMSAA(Canvas* canvas) {
+  return canvas->flags.msaa;
+}
+
+Texture* lovrCanvasGetDepthTexture(Canvas* canvas) {
+  return canvas->depth.texture;
 }
 
 // Buffer
