+++ src/modules/graphics/opengl.h
@@ -11,12 +11,6 @@
 
 #pragma once
 
-#define GPU_CANVAS_FIELDS \
-  bool immortal; \
-  uint32_t framebuffer; \
-  uint32_t resolveBuffer; \
-  uint32_t depthBuffer;
-
 #define GPU_MESH_FIELDS \
   uint32_t vao; \
   uint32_t ibo;
